tracefake/requirements.txt
tensorflow>=2.10.0
streamlit>=1.20.0
opencv-python-headless>=4.7.0
pillow>=9.4.0
exifread>=3.0.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
numpy>=1.23.0
